feat(PB:35): checks extension installation

+++ src/routes/landing/+page.svelte
@@ -1,16 +1,31 @@
 <script>
 	export let data;
 	import { onMount } from 'svelte';
+	import { browser } from '$app/environment';
 	import { PUBLIC_PROLIFIC_LINK } from '$env/static/public';
+	import InstallationPage from '../installation/+page.svelte';
+	import Card from '../../components/card/+page.svelte';
 
 	const keys = Object.values(data);
 	const hasProlificParams = !keys.includes(null);
+	let hasExtension = false;
 
 	onMount(() => {
-		if (hasProlificParams) {
-			localStorage.clear();
-			window.localStorage.setItem('prolific_params', JSON.stringify(data));
-			//check if user has extension
+		if (browser) {
+			if (hasProlificParams) {
+				localStorage.clear();
+				window.localStorage.setItem('prolific_params', JSON.stringify(data));
+			}
+
+			const isInstalled = document.querySelectorAll('[src*="test.js"]');
+			console.log(isInstalled);
+
+			if (isInstalled.length >= 1) {
+				hasExtension = true;
+			} else {
+				hasExtension = false;
+			}
+			console.log(hasExtension);
 		}
 	});
 </script>
@@ -26,4 +41,8 @@
 			</a>
 		</section>
 	</main>
+{:else if !hasExtension}
+	<InstallationPage />
+{:else}
+	<Card h3Value="To be continued..." />
 {/if}
